Restructuration des modèles, de la récupération des vues et de la création des controllers.

+++ app/src/main/java/com/example/makemoremeat/Jeu.kt
@@ -11,7 +11,7 @@ class Jeu : PropertyChangeAware() {
             newValue: Any -> propertyChangeSupport.firePropertyChange(property.name, oldValue, newValue)
     }
 
-    var money: Int by Delegates.observable(0, observer)
+    var money: Long by Delegates.observable(0, observer)
     var moneyPerSecond = 0
     var fastUP = 1 // 1 = x1, 2 = x10, 3 = x25, 4 = xMax
